Fixed tabs.
There was a problem including custom templates as object class in
other ui files so I remove this feature and add a parent tab class
with three boxes (center, left and right) the center box has the expand
property to true. In addition I have set the height_request and
width_request properties in some variables.