Implement select method on search and active_search property on window.
Now we have two methods on search function to select and deselect
the string that we are looking for. It will be usefull to disable
a search whenever we want. In addition active search property now
is quite more complex and it disables a previous search before changing
this property value.